pyproject.toml
[tool.poetry]
name = "youtube-uploader-selenium"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
undetected-chromedriver = "^3.5.3"
loguru = "^0.7.2"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
